+++ src/Utils/ColorGradient.cpp
@@ -28,7 +28,7 @@ Color ColorGradient::getColor(double alpha) const
 	assert(alpha >= 0.0 && alpha <= 1.0);
 
 	Color result;
-	uint64_t index = int(ceil(alpha * totalLength));
+	uint64_t index = uint64_t(ceil(alpha * totalLength));
 
 	for (const ColorGradientSegment& segment : segments)
 	{
